+++ .babelrc
@@ -9,6 +9,6 @@
             "modules":false,
         }]   
         ],
-        "plugins": ["react hot-loader/babel","babel-plugin-transform-class-properties"]
+        "plugins": ["react-hot-loader/babel","babel-plugin-transform-class-properties"]
         //with this plugin we dont have to write constructor inside class synatx
 }
