+++ js/ClientApp.jsx
@@ -1,21 +1,16 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import {BrowserRouter,Route,Switch} from 'react-router-dom'
-import Landing from './Landing'
-import Search from './Search'
+import App from './App'
 
-const App = () =>{
-    return (
-        <BrowserRouter>
-            <div className="app">
-            <Switch>
-            <Route  exact path='/'component={Landing}/>
-            <Route  exact path='/search'component={Search}/>
-            <Route  render= {()=>(<h1>404 page not found</h1>)}/>
-            </Switch>
-            </div>
-        </BrowserRouter>
-    )
+const renderApp = ()=>{
+ReactDOM.render(<App/>,document.getElementById('app'));    
+}
+renderApp();
+
+if(module.hot){
+    module.hot.accept('./App',()=>{
+        renderApp()
+    })
 }
 
-ReactDOM.render(<App/>,document.getElementById('app'));
+//module variable is coming from the webpack thing
