+++ package.json
@@ -56,7 +56,7 @@
     "moxios": "0.4.0",
     "nodemon": "^1.18.10",
     "prettier": "1.3.1",
-    "react-hot-loader": "3.0.0-beta.6",
+    "react-hot-loader": "^4.5.3",
     "react-test-renderer": "^16.8.4",
     "webpack-dev-server": "2.4.5"
   },
