+++ webpack.config.js
@@ -3,15 +3,16 @@ const webpack = require('webpack');
 
 module.exports = {
     context : __dirname,
-    entry : [
-        "react-hot-loader/patch",
-        "webpack-dev-server/client?http://localhost:8080",
-        "webpack/hot/only-dev-server",
-        "./js/ClientApp.jsx"],
+    entry : [//order matter here
+        'react-hot-loader/patch',//HMR
+        'webpack-dev-server/client?http://localhost:8080',//HMR
+        'webpack/hot/only-dev-server',//HMR
+        './js/ClientApp.jsx'],
     devtool:"cheap-eval-source-map", //its very useful coz whenever we make a mistake it will show us our actual code instead of webpack's bundle code
     output:{
         path:path.join(__dirname,'public'),
-        filename:'bundle.js'
+        filename:'bundle.js',
+        publicPath:'/public/'//HMR
     },
     resolve:{
         extensions:['.js','.jsx','.json']
@@ -21,16 +22,17 @@ module.exports = {
         reasons:true,
         chunks:true
     },
+    plugins:[
+        new webpack.HotModuleReplacementPlugin(), //additional abilities to do hot reloading 
+        new webpack.NamedModulesPlugin() // if something goes wrong it will gives you detail about what goes wrong where
+    ],
     devServer:{
-        hot:true,
+        hot:true, //server should be hot //HMR
         publicPath:'/public',
         historyApiFallback:true
         //what does it mean? if you (server) doesnot know about anything just pass it down to client routing
     },
-    plugins:[
-        new webpack.HotModuleReplacementPlugin(),
-        new webpack.NamedModulesPlugin()
-    ],
+   
     module:{
         rules:[
             //if you are using eslint we should use below code before babel-loader
